Add required style
Adjusted required CSS properties for 02-margin-and-padding-2. This
exercise was focused on further exploring how the margin and padding
properties work. Although my first attempt was correct, some minor
improvements were made in the final version after viewing the provided
solution. For example, the top margin of the h1 element was set to 0
pixels to override the default and redundant margins were removed.